+++ eng/mgmt/automation/generate.py
@@ -16,6 +16,12 @@ os.chdir(os.path.abspath(os.path.dirname(sys.argv[0])))
 from parameters import *
 os.chdir(pwd)
 
+# Add two more indent for list in yaml dump
+class ListIndentDumper(yaml.SafeDumper):
+
+    def increase_indent(self, flow=False, indentless=False):
+        return super(ListIndentDumper, self).increase_indent(flow, False)
+
 
 def generate(
     sdk_root: str,
@@ -65,7 +71,7 @@ def generate(
 
     if compile:
         if os.system(
-                'mvn clean verify package -q -f {0}/pom.xml -pl {1}:{2} -am'.
+                'mvn clean verify package -f {0}/pom.xml -pl {1}:{2} -am'.
                 format(sdk_root, GROUP_ID, module)) != 0:
             logging.error('[GENERATE] Maven build fail')
             return False
@@ -148,7 +154,7 @@ def update_service_ci_and_pom(sdk_root: str, service: str):
                 'groupId': GROUP_ID,
                 'safeName': module.replace('-', '')
             })
-            ci_yml_str = yaml.safe_dump(ci_yml, sort_keys = False)
+            ci_yml_str = yaml.dump(ci_yml, sort_keys = False, Dumper = ListIndentDumper)
             ci_yml_str = re.sub('(\n\S)', r'\n\1', ci_yml_str)
 
             with open(ci_yml_file, 'w') as fout:
@@ -382,7 +388,7 @@ def read_api_specs(api_specs_file: str) -> (str, dict):
 def write_api_specs(api_specs_file: str, comment: str, api_specs: dict):
     with open(api_specs_file, 'w') as fout:
         fout.write(comment)
-        fout.write(yaml.safe_dump(api_specs))
+        fout.write(yaml.dump(api_specs, Dumper = ListIndentDumper))
 
 
 def get_and_update_service_from_api_specs(
@@ -539,7 +545,7 @@ def main():
         pwd = os.getcwd()
         try:
             os.chdir(sdk_root)
-            os.system('git add eng/versioning pom.xml {0} {1}'.format(
+            os.system('git add eng/versioning eng/mgmt pom.xml {0} {1}'.format(
                 CI_FILE_FORMAT.format(service),
                 POM_FILE_FORMAT.format(service)))
             os.system(