+++ sdk/spring/azure-spring-boot/src/test/java/com/azure/spring/aad/webapp/AuthorizedClientRepoTest.java
@@ -3,12 +3,12 @@
 
 package com.azure.spring.aad.webapp;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+import org.springframework.boot.test.context.assertj.AssertableWebApplicationContext;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.oauth2.client.OAuth2AuthorizationContext;
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClient;
 import org.springframework.security.oauth2.client.registration.ClientRegistration;
 import org.springframework.security.oauth2.client.web.OAuth2AuthorizedClientRepository;
@@ -16,7 +16,6 @@ import org.springframework.security.oauth2.core.AbstractOAuth2Token;
 import org.springframework.security.oauth2.core.OAuth2AccessToken;
 import org.springframework.security.oauth2.core.OAuth2RefreshToken;
 import org.springframework.security.web.authentication.preauth.PreAuthenticatedAuthenticationToken;
-import org.springframework.test.context.support.TestPropertySourceUtils;
 
 import java.time.Instant;
 import java.util.Optional;
@@ -34,70 +33,76 @@ public class AuthorizedClientRepoTest {
     private MockHttpServletRequest request;
     private MockHttpServletResponse response;
 
-    @BeforeEach
-    public void setup() {
-        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext();
-        TestPropertySourceUtils.addInlinedPropertiesToEnvironment(
-            context,
-            "azure.activedirectory.user-group.allowed-groups = group1, group2",
-            "azure.activedirectory.authorization-server-uri = fake-uri",
-            "azure.activedirectory.tenant-id = fake-tenant-id",
-            "azure.activedirectory.client-id = fake-client-id",
-            "azure.activedirectory.client-secret = fake-client-secret",
-            "azure.activedirectory.authorization.graph.scopes = Calendars.Read"
-        );
-        context.register(AzureActiveDirectoryConfiguration.class);
-        context.refresh();
-
-        AzureClientRegistrationRepository clientRepo = context.getBean(AzureClientRegistrationRepository.class);
-        azure = clientRepo.findByRegistrationId("azure");
-        graph = clientRepo.findByRegistrationId("graph");
-
-        authorizedRepo = new AzureAuthorizedClientRepository(clientRepo);
-        request = new MockHttpServletRequest();
-        response = new MockHttpServletResponse();
-    }
 
     @Test
     public void loadInitAzureAuthzClient() {
-        authorizedRepo.saveAuthorizedClient(
-            createAuthorizedClient(azure),
-            createAuthentication(),
-            request,
-            response);
+        WebApplicationContextRunnerUtils.getContextRunnerWithRequiredProperties()
+            .withPropertyValues(
+                "azure.activedirectory.authorization-clients.graph.scopes = Calendars.Read",
+                "azure.activedirectory.base-uri = fake-uri")
+            .run(context -> {
+                getBeans(context);
 
-        OAuth2AuthorizedClient client = authorizedRepo.loadAuthorizedClient(
-            "graph",
-            createAuthentication(),
-            request);
+                authorizedRepo.saveAuthorizedClient(
+                    createAuthorizedClient(azure),
+                    createAuthentication(),
+                    request,
+                    response);
 
-        assertNotNull(client);
-        assertNotNull(client.getAccessToken());
-        assertNotNull(client.getRefreshToken());
+                OAuth2AuthorizedClient client = authorizedRepo.loadAuthorizedClient(
+                    "graph",
+                    createAuthentication(),
+                    request);
 
-        assertTrue(isTokenExpired(client.getAccessToken()));
-        assertEquals("fake-refresh-token", client.getRefreshToken().getTokenValue());
+                assertNotNull(client);
+                assertNotNull(client.getAccessToken());
+                assertNotNull(client.getRefreshToken());
+
+                assertTrue(isTokenExpired(client.getAccessToken()));
+                assertEquals("fake-refresh-token", client.getRefreshToken().getTokenValue());
+            });
     }
 
     @Test
     public void saveAndLoadAzureAuthzClient() {
-        authorizedRepo.saveAuthorizedClient(
-            createAuthorizedClient(graph),
-            createAuthentication(),
-            request,
-            response);
+        WebApplicationContextRunnerUtils.getContextRunnerWithRequiredProperties()
+            .withPropertyValues(
+                "azure.activedirectory.authorization-clients.graph.scopes = Calendars.Read",
+                "azure.activedirectory.base-uri = fake-uri")
+            .run(context -> {
+                getBeans(context);
 
-        OAuth2AuthorizedClient client = authorizedRepo.loadAuthorizedClient(
-            "graph",
-            createAuthentication(),
-            request);
+                authorizedRepo.saveAuthorizedClient(
+                    createAuthorizedClient(graph),
+                    createAuthentication(),
+                    request,
+                    response);
 
-        assertNotNull(client);
-        assertNotNull(client.getAccessToken());
-        assertNotNull(client.getRefreshToken());
+                OAuth2AuthorizedClient client = authorizedRepo.loadAuthorizedClient(
+                    "graph",
+                    createAuthentication(),
+                    request);
 
-        assertEquals("fake-access-token", client.getAccessToken().getTokenValue());
-        assertEquals("fake-refresh-token", client.getRefreshToken().getTokenValue());
+                assertNotNull(client);
+                assertNotNull(client.getAccessToken());
+                assertNotNull(client.getRefreshToken());
+
+                assertEquals("fake-access-token", client.getAccessToken().getTokenValue());
+                assertEquals("fake-refresh-token", client.getRefreshToken().getTokenValue());
+            });
+    }
+
+    private void getBeans(AssertableWebApplicationContext context) {
+        AADWebAppClientRegistrationRepository clientRepo = context.getBean(AADWebAppClientRegistrationRepository.class);
+        azure = clientRepo.findByRegistrationId("azure");
+        graph = clientRepo.findByRegistrationId("graph");
+
+        authorizedRepo = new AADOAuth2AuthorizedClientRepository(
+            clientRepo,
+            new JacksonHttpSessionOAuth2AuthorizedClientRepository(),
+            OAuth2AuthorizationContext::getAuthorizedClient);
+        request = new MockHttpServletRequest();
+        response = new MockHttpServletResponse();
     }
 
     private OAuth2AuthorizedClient createAuthorizedClient(ClientRegistration client) {
@@ -127,8 +132,8 @@ public class AuthorizedClientRepoTest {
 
     private boolean isTokenExpired(OAuth2AccessToken token) {
         return Optional.ofNullable(token)
-                       .map(AbstractOAuth2Token::getExpiresAt)
-                       .map(expiresAt -> expiresAt.isBefore(Instant.now()))
-                       .orElse(false);
+            .map(AbstractOAuth2Token::getExpiresAt)
+            .map(expiresAt -> expiresAt.isBefore(Instant.now()))
+            .orElse(false);
     }
 }