+++ sdk/spring/azure-spring-cloud-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/servicebus/ServiceBusUtilsTest.java
@@ -14,4 +14,9 @@ public class ServiceBusUtilsTest {
             + "SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=key";
         Assert.assertEquals("namespace", ServiceBusUtils.getNamespace(connectionString));
     }
+
+    @Test
+    public void testGetNamespaceWithInvalidConnectionString() {
+        Assert.assertNull(ServiceBusUtils.getNamespace("fake-connection-str"));
+    }
 }