+++ sdk/cosmos/azure-spring-data-cosmos/CHANGELOG.md
@@ -1,12 +1,22 @@
 ## Release History
 
-## 3.2.0-beta.1 (Unreleased)
+## 3.3.0-beta.1 (Unreleased)
 
+
+### 3.2.0 (2020-12-11)
+#### New Features
+* Updated Spring Data Commons version to 2.3.5.RELEASE.
+* Updated Spring Core version to 5.2.10.RELEASE.
+
+#### Key Bug Fixes
+* Fixed publishing of `spring.factories` file with released jar.
+* Fixed repository query with repeated parameters. 
 
 ### 3.1.0 (2020-10-21)
 #### New Features
 * Added support for `ARRAY_CONTAINS` `CriteriaType`.
 * Updated azure-cosmos version to 4.7.1.
+
 #### Key Bug Fixes
 * Fixed an issue where annotated queries do not pick the annotated container name.